separating OpenAI Plugin base class
